Deep Learning/requirements.txt
streamlit
requests
pillow
langchain
huggingface_hub
openai
